+++ .gitignore
@@ -1,4 +1,5 @@
 build/
 build-win32/
+nxmp/
 nxmp.*
 compile.bat
