+++ include/gui.h
@@ -132,7 +132,7 @@ typedef struct {
 	int playershowcontrols = false;
 	int playershowstats = false;
 	bool masterlock = false;
-	
+	bool clockoc = false;
 	bool first_item;
 	bool focus;
 	int fileHoveredidx = 0;
