+++ source/gui.cpp
@@ -5,11 +5,13 @@
 
 #include "gui.h"
 #include "localfiles.h"
-
 #include "imgui.h"
 #include "imgui_internal.h"
 #include "imgui_impl_sdl.h"
 #include "imgui_impl_opengl3.h"
+#include "SwitchSys.h"
+
+using namespace c2d;
 
 MenuItem item;
 
@@ -44,6 +46,20 @@ namespace GUI {
 		}else{
 			const char *cmd[] = {"show-text", "Masterlock Disabled","2000", NULL};
 			mpv_command_async(libmpv->getHandle(), 0, cmd);
+		}
+		
+	}
+
+	void toggleOC(){
+		item.clockoc = !item.clockoc;
+		if(item.clockoc){
+			const char *cmd[] = {"show-text", "Overclock Enabled","2000", NULL};
+			mpv_command_async(libmpv->getHandle(), 0, cmd);
+			SwitchSys::maxClock();
+		}else{
+			const char *cmd[] = {"show-text", "Overclock Disabled","2000", NULL};
+			mpv_command_async(libmpv->getHandle(), 0, cmd);
+			SwitchSys::defaultClock(SwitchSys::stock_cpu_clock, SwitchSys::stock_gpu_clock, SwitchSys::stock_emc_clock); 
 		}
 		
 	}
@@ -343,6 +359,11 @@ namespace GUI {
 					if (button == SDL_KEY_RSTICK){
 						if(item.state == MENU_STATE_PLAYER){
 							toggleMasterLock();
+						}
+					}
+					if (button == SDL_KEY_LSTICK){
+						if(item.state == MENU_STATE_PLAYER && !item.masterlock){
+							toggleOC();
 						}
 					}
 					if (button == SDL_KEY_MINUS){
